Fix ParticelBatch constructor

+++ src/main/java/net/spell_engine/api/spell/fx/ParticleBatch.java
@@ -72,7 +72,7 @@ public class ParticleBatch { public ParticleBatch() { }
 
     // Compatibility constructors
 
-    public ParticleBatch(String particle_id, Shape shape, Origin origin, int count, float min_speed, float max_speed) {
+    public ParticleBatch(String particle_id, Shape shape, Origin origin, float count, float min_speed, float max_speed) {
         this(particle_id, shape, origin, null, 0, 0, count, min_speed, max_speed, 0, 0, 0, false);
     }
